refactor(#113926): Extracting "JsonValueOfJsonString" to separate file.
